Changed the data type of the column  to BIGINT

+++ authorization_server/src/main/resources/db/migration/V1__create_all_tables.sql
@@ -43,7 +43,7 @@ CREATE TABLE token_settings
 (
     id                    INT AUTO_INCREMENT
         PRIMARY KEY,
-    time_to_live_in_hours INT         NOT NULL,
+    time_to_live_in_hours BIGINT         NOT NULL,
     format                VARCHAR(50) NOT NULL,
     client_id             INT         NOT NULL,
     CONSTRAINT token_settings_clients_id_fk
